app/requirements.txt
aiofiles==0.5.0
aiohttp==3.7.0
PyYAML==5.3.1
